requirements.txt
git+https://github.com/shahbuland/PerceptualSimilarity
torchmetrics
einops
ema-pytorch
vector-quantize-pytorch
ray
omegaconf
wandb
kornia
datasets
huggingface-hub
python-dotenv
alias-free-torch
git+https://github.com/sypsyp97/convnext_perceptual_loss
wandb[media]
rotary-embedding-torch
boto3
diffusers
pytorch-wavelets